requirements.txt
seatable-api
openpyxl
pandas
python-dotenv